+++ src/gui/drawables/animatedstatelistdrawable.cc
@@ -245,7 +245,7 @@ int AnimatedStateListDrawable::parseItem(XmlPullParser&parser,const AttributeSet
         while ((type = parser.next()) == XmlPullParser::TEXT) {
         }
         if (type != XmlPullParser::START_TAG) {
-            throw std::logic_error(//parser.getPositionDescription()
+            throw std::logic_error(parser.getPositionDescription()+
                     ": <item> tag requires a 'drawable' attribute or child tag defining a drawable");
         }
         dr = Drawable::createFromXmlInner(parser,atts);
@@ -268,7 +268,7 @@ int AnimatedStateListDrawable::parseTransition(XmlPullParser&parser,const Attrib
         while ((type = parser.next()) == XmlPullParser::TEXT) {
         }
         if (type != XmlPullParser::START_TAG) {
-            throw std::logic_error(//parser.getPositionDescription()
+            throw std::logic_error(parser.getPositionDescription()+
                             ": <transition> tag requires a 'drawable' attribute or child tag defining a drawable");
         }
         dr = Drawable::createFromXmlInner(parser, atts);