+++ src/gui/view/layoutinflater.h
@@ -32,9 +32,9 @@ private:
     void parseInclude(XmlPullParser&parser, Context*,View*prent,const AttributeSet& attrs);
 protected:
     View* createView(const std::string& name, const std::string& prefix,const AttributeSet& attrs);
-    View* createViewFromTag(View* parent,const std::string& name, Context* context,const AttributeSet& attrs,bool ignoreThemeAttr);
-    void rInflateChildren(XmlPullParser& parser, View* parent,const AttributeSet& attrs,bool finishInflate);
-    void rInflate(XmlPullParser& parser, View* parent, Context* context,const AttributeSet& attrs, bool finishInflate);
+    View* createViewFromTag(View* parent,const std::string& name, Context* context,AttributeSet& attrs,bool ignoreThemeAttr);
+    void rInflateChildren(XmlPullParser& parser, View* parent,AttributeSet& attrs,bool finishInflate);
+    void rInflate(XmlPullParser& parser, View* parent, Context* context,AttributeSet& attrs, bool finishInflate);
 public:
     static LayoutInflater*from(Context*context);
     static ViewInflater getInflater(const std::string&);